refactor(sections): extract the leading-paragraph readers into preamble.rs
Pure motion: where prose ends and structure begins, which paragraphs are
banners rather than a description, and which are the tool's own
option-error complaint about the --help probe.
